move deps from contracts to commons
