+++ packages/common/assets/verification_key.json
