+++ packages/common/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@hojicha/common",
+  "version": "0.0.1",
+  "description": "Common utils for frontend and backend",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
