+++ packages/common/src/externalNullifier.js
