--- packages/contracts/lib/index.js
@@ -1,1 +1,0 @@
-const { compileContracts } = require('./compile')
